requirements.txt
borb==2.1.10
flask-login==0.6.2
flask==2.2.3
jinja2==3.1.2
libsass==0.22.0
markupsafe==2.1.2
mollie-api-python==2.13.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pyvat==1.3.17
sendgrid==6.10.0
sqlalchemy==2.0.7
werkzeug==2.2.3